requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.2
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise==6.11.0
